pyproject.toml
[tool.poetry]
name = "ae-de-dbt-demo"
version = "0.1.0"
description = "Mini AE to DE demo using Poetry, Postgres, dbt and fake ingestion"
authors = ["Armando Rodrigues"]
readme = "README.md"
packages = [{ include = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
faker = "^37.1.0"
dbt-postgres = "^1.9.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
python-dateutil = "^2.9.0.post0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"